requirements.txt
flask
Pillow
tensorflow
opencv-python
matplotlib
